requirements.txt
fastapi==0.135.1
uvicorn[standard]==0.34.0
pydantic==2.12.0
pydantic-settings==2.8.0
google-genai==1.65.0
python-multipart==0.0.20
# pydantic 2.12.0 requires typing-extensions>=4.14.1
typing-extensions>=4.14.1
python-dotenv==0.21.0